feat: device menu (#89)
* feat: device menu

We need an intuitive way to load and save configuration files. Right
now, the only option provided by the GUI is the buttons on the
GroupPresetTableWidget, which might not be open.

This changeset provides a solution using a new Menu (titled "Device"),
aligning with the existing mmstudio solution.

* Window menu: Ignore actions in other menus

* Move and add ellipses to menu actions

* Rename device menu

It's "Devices" now

* Extract Save/Load modal widgets

* Configure widgets via __init__

This fixes the tests :)