fix: pre-load MSVC DLLs to block old versions (#94)
* fix: pre-load MSVC DLLs to block old versions

Before importing any third-party packages, load these DLLs from the
system-installed copies (or copies bundled by the Python interpreter in
the case of `vcruntime140[_1].dll`).

* Suppress exceptions when pre-loading MSVC DLLs

If there is a failure here, it means that the VC Redistributables are
not installed on the system. That probably means Micro-Manager has not
been installed, in which case this whole fix is (probably) not needed.